S19 - Project 10 - Automatic Ticket Classification Tool/Automatic Ticket Classification Tool/requirements.txt
langchain
streamlit
openai
tiktoken
python-dotenv
pinecone-client
pypdf
joblib
pandas